(model) Leen's complete model: a stable version to be able to decode PCs from the input neurons' activities.
Why
The reason of this fix is aligned to the research goal: the first stage of my research goal is to build a biological realistic model to capture PCs from the neurons' activities.

What changed
I tested different learning rates for different types of the input matrices $X$ with varied eigenvalues.

A conclusion is that the learning rate should be adjusted by the statistical properties of the input matrices (i.e., eigenvalues).

Data / Code context
In this version, I used a case that there are 10 input neurons, but only 3 strongest PCS.

Params
This information is shown in the Juypter Notebook.

Results

How to use

Links

Comments